CloudFunctions/SendGrid/SendGridSaveEventWebhook/requirements.txt
sendgrid
google-cloud-bigquery
google-cloud-pubsub
google-cloud-logging

